Refactored url parser to a different file
